backend/requirements.txt
fastapi
python-multipart
uvicorn
pydantic
langchain==0.3.28
langchain-groq
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
slack_sdk
notion-client
PyGithub
requests
pypdf
